backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "thhsie",email = "teshiii@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"